requirements.txt
# ETL Pipeline Dependencies

# Google Cloud & Drive
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
gspread==6.0.0

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Additional utilities
python-dateutil==2.8.2
pytz==2024.1
